Extract hooks to files
