+++ packages/ui/src/working-groups/hooks/useWorkingGroup.ts
@@ -1,28 +1,8 @@
-import {
-  useCountWorkingGroupOpeningsQuery,
-  useCountWorkingGroupWorkersQuery,
-  useGetWorkingGroupQuery,
-} from '../queries'
-import { asWorkingGroup, WorkerStatusTypename } from '../types'
+import { useGetWorkingGroupQuery } from '../queries'
+import { asWorkingGroup } from '../types'
 
 interface WhereInput {
   name?: string | undefined
-}
-
-export function useCountOpenings(groupId: string) {
-  const { data, loading } = useCountWorkingGroupOpeningsQuery({
-    variables: { groupId_eq: groupId, status_json: { isTypeOf_eq: 'OpeningStatusOpen' } },
-  })
-
-  return { isLoading: loading, openings: data?.workingGroupOpeningsConnection.totalCount }
-}
-
-export function useCountWorkers(groupId: string) {
-  const { data, loading } = useCountWorkingGroupWorkersQuery({
-    variables: { groupId_eq: groupId, status_json: { isTypeOf_eq: WorkerStatusTypename['active'] } },
-  })
-
-  return { isLoading: loading, workers: data?.workersConnection.totalCount }
 }
 
 export function useWorkingGroup(where: WhereInput) {
